+++ library/createDirectory.js
@@ -1,6 +1,6 @@
 // Node modules.
-const { exists, lstat, mkdir } = require(`fs`),
-	{ join, sep } = require(`path`);
+const fs = require(`fs`),
+	path = require(`path`);
 
 /**
  * Create a directory at the path.
@@ -8,21 +8,21 @@ const { exists, lstat, mkdir } = require(`fs`),
  */
 module.exports = function(directory) {
 	// Split directory path into sequential directory name segments.
-	directory = directory.split(sep);
+	directory = directory.split(path.sep);
 	// Iterate over each part of the directory path and add the next part to each subsequent promise.
 	return directory.reduce(function(promise, directory) {
 		return promise.then(function(previous) {
-			directory = join(previous, directory);
+			directory = path.join(previous, directory);
 			// Check if directory already exists.
 			return new Promise(function(resolve) {
-				exists(directory, function(exists) {
+				fs.exists(directory, function(exists) {
 					resolve(exists);
 				});
 			}).then(function(exists) {
 				if (exists) {
 					// Directory already exists go to next element of path.
 					return new Promise(function(resolve, reject) {
-						lstat(directory, function(error, stats) {
+						fs.lstat(directory, function(error, stats) {
 							if (error) {
 								return reject(error);
 							}
@@ -36,7 +36,7 @@ module.exports = function(directory) {
 				}
 				// Create the directory.
 				return new Promise(function(resolve, reject) {
-					mkdir(directory, function(error) {
+					fs.mkdir(directory, function(error) {
 						if (error && error.code !== `EEXIST`) {
 							return reject(error);
 						}