+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
+## 1.0.0-beta.0 (2018-09-19)
+### Changed
+- Restructured project files.
+- Update examples to match with new versions of any modules.
+
 ## 0.2.0 (2018-09-11)
 ### Added
+- `CHANGELOG.md` added.
 - `scanDirectory` function added to helper functions.
 - `remove` function added to helper functions.
 - `writeFiles` function added to helper functions.